Add rocketmap installation with map config
Adds an install check and prompts to download and install RocketMap (if
needed) and write the map config before clearing the prompt.

One caveat is that it needs to be running as root, which may not go down
so well once it's on a real server. Theoretically NOPASSWD options
should be viable in `sudo visudo` but didn't seem to help on localhost.