Switch to bash script for restarting areas
For some reason artisan's a dick and so `ps aux` returns empty when the
artisan schudule is ran from cron, meaning nothing can be restarted.

This moves the critical parts to a shell script, with artisan now only
generating a list of search strings, from which the script can find the
PIDs using `ps` natively before stopping and restarting.

Since the restarting is no longer happening in Laravel (at least when
done automatically) we also now have a new command to log restarts.