Add Discord config page
Moves Discord's bot token and server/guild ID options to a new page in
the admin panel rather than having them hard-coded in env vars.

Adds a new wrapper for RestCord\DiscordClient so we don't have to keep
loading up the config and rewriting the code everywhere we use Discord.