Fix discord config migration
For whatever reason, this gets instantiated by the DiscordPrep command
even when it hasn't specifically been called. Maybe we can load Discord
from some other init-type method instead of __construct() to avoid this.

For now, wrapping the db-stuff that needs potentially-non-existent
tables in a try/catch block is a good enough workaround.