Fix team colour bug
Despite the array cast, colours is somehow ending up as an object and so
shit breaks when they're set... or something. Shit's confusing with the
getColoursAttribute method added into the mix.

Changing the usages from [$foo] to ->$foo seems to be enough, but for
good measure we'll cast arrays to objects when we're getting colours
too. Theoretically this should only happen when colours aren't set.